refactor: make dedicated `utils` folder in `components`
