+++ components/events/event-card.tsx
@@ -2,8 +2,8 @@ import Image from "next/image";
 import { BiTimeFive } from "react-icons/bi";
 import sameDay from "../../lib/same-day";
 import { Event } from "../../types/event";
-import FormattedDate from "../formatted-date";
-import FormattedTime from "../formatted-time";
+import FormattedDate from "../utils/formatted-date";
+import FormattedTime from "../utils/formatted-time";
 import { eventColors } from "../../lib/event-colors";
 import shorten from "../../lib/shorten";
 import { motion } from "framer-motion";
